requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
influxdb-client==1.33.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pipreqs==0.4.11
plotly==5.10.0
protobuf==3.20.1
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
reactivex==4.0.4
requests==2.28.1
rich==12.5.1
schedule==1.1.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.12.0
streamlit-nested-layout==0.1.1
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
yarg==0.1.9
zipp==3.8.1
python-dotenv==1.0.0
